requirements/dev-req.txt
appdirs==1.4.4
asgiref==3.5.0
astroid==2.9.3
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
Django==4.0.3
django-extensions==3.1.5
django-money==2.1.1
djangorestframework==3.13.1
drf-spectacular==0.21.2
inflection==0.5.1
isort==5.10.1
jedi==0.18.1
jsonschema==4.4.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
parso==0.8.3
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
ptpython==3.0.20
py-moneyed==1.2
pycodestyle==2.8.0
Pygments==2.11.2
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
wcwidth==0.2.5
wrapt==1.13.3
